[net] Add documentation for feature process & supported platforms
As //net is used in a variety of "Not Chrome (the Browser)"
cases, add documentation to //net/docs to help enumerate the
various supported projects using //net and their capabilities
as it relates to feature design/support.

In addition, document a bit more about how to a //net feature
lives, including common design patterns that tend to trip up
folks proposing features for //net. This is just an initial
draft to be iterated and refined on.

BUG=none
TEST=none

Review-Url: https://codereview.chromium.org/2700673002
Cr-Original-Commit-Position: refs/heads/master@{#451795}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7390e5594fb6f8468359b519580dacb29fdf269b