Fix BufferedSpdyFrameTest.OnAltSvc.
The mock ALTSVC frame in BufferedSpdyFrameTest.OnAltSvc was not spec compliant
with its non-zero stream id and non-empty origin value.  While this is of no
practical importance, because the framing layer this test is exercising does
not enforce this part of the specification, it is good practice to use spec
compliant data in tests.  This CL splits the test into two: one with
stream id 0 and non-empty origin, and another one with stream id other than
zero and empty origin, both combinations conforming to the specs.

BUG=696653

Review-Url: https://codereview.chromium.org/2718053003
Cr-Original-Commit-Position: refs/heads/master@{#454137}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ce54af31e8cf2b9a1da25a30b327087afc0862ef